requirements.txt
flask
google-generativeai
sentence-transformers
qdrant-client
python-dotenv